+++ docs/source/changelog.rst
@@ -1,4 +1,4 @@
-.. Created by changelog.py at 2025-02-04, command
+.. Created by changelog.py at 2025-02-28, command
    '/Users/giffler/.cache/pre-commit/repoecmh3ah8/py_env-python3.13/bin/changelog docs/source/changes compile --categories Added Changed Fixed Security Deprecated --output=docs/source/changelog.rst'
    based on the format of 'https://keepachangelog.com/'
 
@@ -6,8 +6,14 @@
 CHANGELOG
 #########
 
-[Unreleased] - 2025-02-04
+[Unreleased] - 2025-02-28
 =========================
+
+Added
+-----
+
+* Add new duping SSH executor to support HPC ssh command keys
+* Added Auditor accounting plugin
 
 Fixed
 -----